refactor: DAO 클래스에 대한 Fake 클래스 생성
