+++ src/test/java/domain/JanggiBoardTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.dao.MemoryPieceDao;
 import domain.piece.Piece;
 import domain.piece.PieceType;
 import domain.piece.Position;
@@ -26,8 +27,7 @@ public class JanggiBoardTest {
         Piece pawn2 = new Piece(Team.HAN, PieceType.PAWN, new Position(5, 1));
         afterBoard.add(pawn2);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(4, 1);
         Position targetPosition = new Position(5, 1);
@@ -53,8 +53,7 @@ public class JanggiBoardTest {
         Piece choCha2 = new Piece(Team.CHO, PieceType.CHA, new Position(8, 1));
         afterBoard.add(choCha2);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(4, 1);
         Position targetPosition = new Position(8, 1);
@@ -76,9 +75,7 @@ public class JanggiBoardTest {
         beforeBoard.add(choCha1);
         beforeBoard.add(choCha2);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
-
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
         Position startPosition = new Position(4, 1);
         Position targetPosition = new Position(8, 1);
 
@@ -96,8 +93,7 @@ public class JanggiBoardTest {
         List<Piece> beforeBoard = new ArrayList<>();
         beforeBoard.add(choPo);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(4, 1);
         Position targetPosition = new Position(8, 1);
@@ -118,8 +114,7 @@ public class JanggiBoardTest {
         beforeBoard.add(choCha);
         beforeBoard.add(choPawn);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(1, 1);
         Position targetPosition = new Position(8, 1);
@@ -140,8 +135,7 @@ public class JanggiBoardTest {
         beforeBoard.add(choPo1);
         beforeBoard.add(choPo2);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(8, 2);
         Position targetPosition = new Position(8, 9);
@@ -164,8 +158,7 @@ public class JanggiBoardTest {
         beforeBoard.add(choGung);
         beforeBoard.add(choPo2);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(8, 2);
         Position targetPosition = new Position(8, 8);
@@ -191,8 +184,7 @@ public class JanggiBoardTest {
         afterBoard.add(choPo2);
         afterBoard.add(choGung);
 
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(beforeBoard);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(beforeBoard));
 
         Position startPosition = new Position(8, 2);
         Position targetPosition = new Position(8, 8);
@@ -213,8 +205,7 @@ public class JanggiBoardTest {
         //given
         List<Piece> board = new ArrayList<>();
         board.add(new Piece(gungTeam, PieceType.GUNG, new Position(1, 1)));
-        FakeBoardGenerator boardGenerator = new FakeBoardGenerator(board);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(boardGenerator, 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(board));
 
         // when
         boolean actual = janggiBoard.existGung(team);
@@ -225,7 +216,7 @@ public class JanggiBoardTest {
     @Test
     void 시작_위치에_기물이_존재하지_않는_경우_예외를_발생시킨다() {
         List<Piece> board = new ArrayList<>();
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(new FakeBoardGenerator(board), 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(board));
 
         assertThatThrownBy(() -> janggiBoard.move(new Position(1, 1), new Position(1, 2)))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -237,7 +228,7 @@ public class JanggiBoardTest {
         List<Piece> board = new ArrayList<>();
         Piece pawn = new Piece(Team.CHO, PieceType.PAWN, new Position(1, 1));
         board.add(pawn);
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(new FakeBoardGenerator(board), 1L);
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(board));
 
         Piece selectedPiece = janggiBoard.findSelectedPiece(new Position(1, 1));
 
@@ -247,8 +238,7 @@ public class JanggiBoardTest {
     @Test
     void 시작_위치가_기물이_없는_위치라면_예외를_발생시킨다() {
         List<Piece> board = new ArrayList<>();
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(new FakeBoardGenerator(board), 1L);
-
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(board));
         assertThatThrownBy(() -> janggiBoard.findSelectedPiece(new Position(1, 1)))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("기물이 존재하지 않는 위치입니다.");
@@ -268,8 +258,7 @@ public class JanggiBoardTest {
         board.add(new Piece(Team.HAN, PieceType.GUNG, new Position(2, 2)));
         board.add(new Piece(Team.HAN, PieceType.SANG, new Position(2, 3)));
 
-        JanggiBoard janggiBoard = JanggiBoard.initBoard(new FakeBoardGenerator(board), 1L);
-
+        JanggiBoard janggiBoard = new JanggiBoard(new MemoryPieceDao(board));
         int scoreSum = janggiBoard.calculateTeamScore(team);
 
         assertThat(scoreSum).isEqualTo(score);
